requirements.txt
alembic==0.9.6
aniso8601==1.3.0
bcrypt==3.1.4
cffi==1.11.2
click==6.7
coverage==4.4.2
enum-compat==0.0.2
eventlet==0.21.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SocketIO==2.9.3
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gem==0.1.12
greenlet==0.4.15
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10.1
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
Pillow==7.1.2
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
python-engineio==2.0.2
python-socketio==1.8.4
pytz==2017.3
selenium==3.8.1
six==1.11.0
SQLAlchemy==1.2.0
SQLAlchemy-ImageAttach==1.1.0
Wand==0.6.1
Werkzeug==0.15.3
